Add `pattern` module with stricter transcript recording (#62)
This adds a new `pattern` module that contains a representation of an
interactive interaction pattern together with recording and playback
states. The pattern contains types, labels and hierarchical grouping of
(sub)protocols.

As discussed with @WizardOfMenlo, this only adds the pattern module for
review and does not use it anywhere yet. The intention is for a follow
up PR to use it as a substitute for `DomainSeparator`.

---------

Co-authored-by: Giacomo Fenzi <giacomofenzi@outlook.com>